feat: replace SavedAddresses with ShippingAddresses
